hisi_riscv_hal::i2c

Struct I2c

pub struct I2c<'d, T> { /* private fields */ }

I2C master driver bound to instance T (I2c0/I2c1).

Implementations

impl<'d> I2c<'d, I2c0<'d>>

pub fn new_i2c0(_i2c: I2c0<'d>, speed: Speed) -> Self

Create and configure the I2C0 master at the given bus Speed.

impl<'d> I2c<'d, I2c1<'d>>

pub fn new_i2c1(_i2c: I2c1<'d>, speed: Speed) -> Self

Create and configure the I2C1 master at the given bus Speed.

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

impl<T> I2c<'_, T>

pub fn write(&mut self, addr: u8, data: &[u8]) -> Result<(), I2cError>

Write data to the 7-bit addr (START, address+W, bytes, STOP).

pub fn read(&mut self, addr: u8, buf: &mut [u8]) -> Result<(), I2cError>

Read buf.len() bytes from the 7-bit addr (START, address+R, bytes with NACK on the last, STOP).

pub fn write_read( &mut self, addr: u8, wr_buf: &[u8], rd_buf: &mut [u8], ) -> Result<(), I2cError>

Combined write-then-read with repeated START (Sr) between operations, matching the I2C specification for register-based device access.